Refactor Complex module for finer granularity.
