+++ Sources/ComplexModule/Complex+AdditiveArithmetic.swift
@@ -1,0 +1,42 @@
+//===--- Complex+AdditiveArithmetic.swift ---------------------*- swift -*-===//
+//
+// This source file is part of the Swift Numerics open source project
+//
+// Copyright (c) 2019-2021 Apple Inc. and the Swift Numerics project authors
+// Licensed under Apache License v2.0 with Runtime Library Exception
+//
+// See https://swift.org/LICENSE.txt for license information
+//
+//===----------------------------------------------------------------------===//
+
+import RealModule
+
+extension Complex: AdditiveArithmetic {
+  /// The additive identity, with real and imaginary parts both zero.
+  ///
+  /// See also: `one`, `i`, `infinity`
+  @_transparent
+  public static var zero: Complex {
+    Complex(0, 0)
+  }
+  
+  @_transparent
+  public static func +(z: Complex, w: Complex) -> Complex {
+    return Complex(z.x + w.x, z.y + w.y)
+  }
+  
+  @_transparent
+  public static func -(z: Complex, w: Complex) -> Complex {
+    return Complex(z.x - w.x, z.y - w.y)
+  }
+  
+  @_transparent
+  public static func +=(z: inout Complex, w: Complex) {
+    z = z + w
+  }
+  
+  @_transparent
+  public static func -=(z: inout Complex, w: Complex) {
+    z = z - w
+  }
+}
