+++ Sources/ComplexModule/Complex+Differentiable.swift
@@ -2,7 +2,7 @@
 //
 // This source file is part of the Swift Numerics open source project
 //
-// Copyright (c) 2019 - 2020 Apple Inc. and the Swift Numerics project authors
+// Copyright (c) 2019 - 2021 Apple Inc. and the Swift Numerics project authors
 // Licensed under Apache License v2.0 with Runtime Library Exception
 //
 // See https://swift.org/LICENSE.txt for license information
