+++ Sources/ComplexModule/Complex+ElementaryFunctions.swift
@@ -1,4 +1,4 @@
-//===--- ElementaryFunctions.swift ----------------------------*- swift -*-===//
+//===--- Complex+ElementaryFunctions.swift --------------------*- swift -*-===//
 //
 // This source file is part of the Swift.org open source project
 //
